17332 Fix pushed MKPs in distributed monitoring being deleted on update
The logic to find installed MKPs takes pushed MKPs into account which
would only exist in local_enabled_packages_dir. On update, installation
of these MKPs would succeed, adding their manifest to the installer. Due
to Werk 14655, we remove the "previously enabled package", however this
deletes the original pushed copy of the MKP from
local_enabled_packages_dir.

This causes a problem as, the update process tries to link the MKP into
local_enabled_packages_dir and fails, leaving the application in a state
where it lists the MKP as enabled (manifest is installed) but is unable
to disable it (nothing linked in local_enabled_packages_dir).

SUP-19173

Change-Id: Iaca992ce86ccea643f601bdb684224217ff90520