Don't even start auto discovery job in case there is no work to do
So far we started the background job every 5 minutes, which would create
a new process, spin up the UI envirionment to trigger a local automation
which would start a new subprocess that spins up the base environment.
Just to exit immediately because there is no work to do.

In many sites the auto discovery is not configured at all, so there will
never be work for the job. Even if it is configured, there will be no
work for the job most of the time.

Change-Id: Ic1b610aa9f1bb537cb6462278221ee3f42eed254