17333 FIX dropdowns in grouped dicts breaking scrolling
When editing a rule based on a form spec which contains a dictionary
with grouped elements that contain a dropdown (the SingleChoice form
spec for example), the page would no longer be scrollable after opening
one such dropdown.

This lead to the `httpv2` rule breaking the scrolling behavior in
multiple locations.

With this werk, we fix the underlying issue when rendering a dropdown
in the browser. Now, the page will continue to be scrollable after such
a dropdown has been clicked.

SUP-19467

Change-Id: I6ce2b2d740bfc6468cc3baecedefcf2fa9c35c76