testlib.licensing: Refactor fixture to function
Instead of defining the fixture "fixture_licensing_key_or_cert" in the
testlib and then importing it in some conftest files, this change will
refactor the fixture into a function which is then imported and used
in a fixture. It is a bad practice to import fixtures and comes with
several problems (including import and typing errors).

Change-Id: Ia9e0957c947f37dd6e86962a9894816b618b256b