17390 FIX Fix update of graph collections with combined graphs
In earlier Checkmk versions, there was some update logic for visual contexts.
This was dropped in the belief that all visual contexts had been udpated.
However, the visual contexts of combined graphs were forgotten. We simply copy
over the update logic from 2.2, which is the newest version that still contains
it.

SUP-21001

Change-Id: Ie469a889c4cfeb53e95307137437df08de84cb30