precompiled host checks: make plugin loading interruptable
To clean up the whole plugin loading we need to make up our
mind if plugin loading should be interruptable by SIGINT.
Prior to this change the loading of the new-style plugins
was not interruptable, but the loading of the legacy
plugins was. We assume all loading should be interruptable.

Change-Id: I0f3372c252920455bada41f2fcaf86bd4662f3fe