17334 Catch invalid "Explicit hosts" conditions
When invalid host names were passed into the "Explicit hosts" condition
while creating or editing a rule, Checkmk would crash while trying to
render the host condition. The crash is caused by a __new__ call to
`HostName`. We now validate the non-regex input using the same
validator.

With this werk, the form will warn you that you have entered an invalid
Explicit hosts condition. This allows you to fix the error without
losing the rest of the entered inputs.

CMK-19405

Change-Id: I2ab8f8850c6e8b671d6abc513d98d762cc2fffa0