notifications: remove default service event option
Remove status change from Any to UNKN as a default option. This is not a
common case that a user would be notified on. This type of status change
is usually because a service item was removed (filesystem, network
interface, etc.), which usually is expected and would be dealt with by
the monitoring admin.

CMK-19895

Change-Id: I1419c680cb8a6bde3adf4777d63f2ee5b393bc75