Sorters: Get rid of useless dependencies
It seems that the arguments of the Sorter was introduced to have the
same signature as the Painter. This is not really helpful and pulls a
lot of things into the class hierarchy. We now (temporarily) get rid of
all constructor arguments and hand over the dependencies directly to the
methods requiring them.

Change-Id: Ib46b9e64a00e4f154ef837178db475bb6b6dba08