Fixed partial filter
The data for "host_down" and "service_period" has not yet been set for the
rows, so a filter should not try to access those columns.  Although this is
a correctness fix, it doesn't make any difference for our own SW, which
doesn't use a filter on these fields for the "statehist" table.  At least I
think so...  :-}

Be a bit more explicit about the "current_*" columns, just for readability.

Change-Id: I31d5b21176a0ca9c525cc960b27f5b19c870eec1