Prepare import cycle fix: notif param registry>watolib
The legacy converter accesses wato for numerous legacy wato groups.
It itself is accessed by the notification_parameter_registry which is
accessed by the rest api.

This commit is step one from stopping watolib from importing wato at
import time.

The fix is suboptimal though might be acceptable as the legacy converter
can soon be removed and the alternative would be a major refactoring to
pull out the wato groups out of wato.

CMK-19751

Change-Id: Ia0e2a2661ae204790a570872a99737e45258f8b0