notifications: disallow deleting parameter with rules
We want to prevent users from making preventable errors. In this change,
we are preventing users from deleting notification parameters that are
being used in existing notification rules.

Instead of deleting the parameter, an error message is raised in the GUI
along with anchor links to the rules that are currently using the
notification parameter.

CMK-19149

Change-Id: I40dd135459651c0410c8bd04060dab87ceb8765d