painter: decode semi-colon from monitoring history
The plugin output and comments are saved in a Nagios-compatible format
(semi-colon-separated). Therefore, it is necessary to escape semi-colons
to `%3B` when writing out history.

In the GUI, where plugin output or comments from the monitoring history
are displayed, we can translate `%3B` back to a semi-colon using a
simple replace.

Although this is not 100% correct (`%3B` could have been in the original
text), it will probably do the right thing in 99.9% of use cases and is
therefore a valid fix for now.

SUP-20926

Change-Id: Iae15c50b9c9dcf06913a65001818e11710852ceb