agent based plugins: do not add created mgmt plug-ins
For services created from the "management board as host attribute"
setup, management versions of the plug-ins are created on the fly.
Let's not add them to the set of registered plug-ins, to at least
maintain a _consistent_ globale state.
This setup is discouraged and deprecated anyway, and the global
state is being removed.

Change-Id: I6c818aa4c5922de7b6f70fa7c44ef4286a8fb9cc