gui: improve title of single choice editable component
It is inadequate to have a generic "Add parameter" as the slide out
title. It is important that the user knows which entity they have
selected and are currently creating a parameter for, e.g. "Add HTML
Email parameter".

In order for the component to remain generic, an additional helper was
added to fetch human-readable titles relevant to notification parameters.

CMK-19990

Change-Id: I7840a168d2542ebb640b7604013725dc74ab9d56