agent_based: Refactor get_relevant_raw_sections
This refactors (renames, and moves) the function
to avoid accessing the global state by taking
the plugins themselves (instead of their names)
as arguments.

As a result, we no longer have to distinguish
check and inventory plugins. In particular the
callsite in cmk.base.core_nagios can be simpified
significantly as a result of that.

Change-Id: Ib5a16a926270c9ceefce1070d8c42907c1e5ac11