17356 FIX Decode semi-colon from monitoring history
The plugin output and comments are saved in a Nagios-compatible format
(semi-colon-separated). Therefore, it is necessary to escape semi-colons
to `%3B` when writing out history.

In the GUI, where plugin output or comments from the monitoring history
are displayed, we can translate `%3B` back to a semi-colon using a
simple replace.

Change-Id: Ib8d82f9de51a08692816cdc891a4505bd1d950fb