docker tests: Refactor checkmk container usage
The checkmk container is now wrapped into a class
which contains all the attributes and methods for
handling the container. This class also has an openapi
attribute to easily access the API of the Checkmk site
in the container.

NOTE: The CSE auth provider setup is dropped with this
change (as it did not currently work anyway).

CMK-20255

Change-Id: I4c803b142e8332755b370f0035d169262f15fbdc