vue: add minimal button variant
The minimal button variant should be used if you want to differentiate
between more prominent buttons like primary and secondary. Unlike the
tertiary variant (also used for differentiation), the minimal button has
subtle styling, but enough to show that it is indeed clickable. It
accomplishes this by changing its text color on hover.

Change-Id: I0a3735efa8a884ad771030dd5594abf7d627346f