15347 FIX mk-job: currently running job was not correctly reported
`mk-job` creates multiple files: a stat file for finished jobs, and a running
file for currently running jobs. The `check_mk_agent` then collects those files.
Previously to this change, the sorting of those files was important.

The first file belonging to a certain job determined the state of the aggregated job.

Now the aggregated job will be set to running although the running file is not
at first position.

SUP-21221

Change-Id: Ib2cd1d8681c68691d491016bc82541356fda4312