Use upstream_build in build-cmk-packages
This saves time during a release as the windows agent artifacts are not
built several times. upstream_build will wait for the completion of the
first build with the same parameters as long a job is already building and
not waiting in the build queue.
Previously no build with the release version as VERSION was found and thereby
built, by every edition. Leading to very long runs and delays

Change-Id: Ib1aea64e40b6c271e95446e5bb37be36b48d9c0e
(cherry picked from commit 17642d23c0dc3c69f455c8c4e106013ad3feddd8)