testlib.docker: Download CMK package before test
So far, when running the docker test locally, one had to
manually download the Checkmk package to the docker_image
subfolder.

This change will improve that behaviour: In case the checkmk
package is not found in the docker_images folder, it will be
downloaded to be able to proceed with the test.

CMK-20255

Change-Id: Ia9ff3c4b545d1452082d7bf67bca7ed69ffd4774