17405 Support Diagnostics: Checkmk server host parameter
This werk introduces a parameter "Checkmk server host".
In many cases, the support diagnostics can automatically determine the
host, that represents the Checkmk server. If that's not possible, the
user can now manually select that host before creating the dump.

This information is needed for the Performance Graphs and the Checkmk
Overview.

Change-Id: I954ca885e29cafb93667ce5c20789c1e167a0c44