omd: run tests on changes in dependencies II
omdlib relies on `cmk/utils` for some its functionality. This means some
of the omd unit tests may break, if `cmk/utils` changes. The tests now
run, if a change in these packages is observed.

Since `cmk/utils` is not a proper package yet, we will add `cmk/utils`
to the omd files. This should be cleaned up at some point.

CMK-19973

Change-Id: I4b4b4a65b112a0d7a1078619fcea9d5b55a7545d