Remove dependency on cmk.fetchers from cmk.gui
Pulling in cmk.fetchers just for this class is a huge overkill in terms
of memory consumption. The problem is that cmk.fetchers pulls in netsnmp
and pyghmi, which consume memory even if they are not used in this
context. It is better to accept a bit of code duplication in this case.

Change-Id: I122a0238c1190de63b5b2bd085904e3a0c346d57