Refactor job registration to a plugin registry
This is a refactoring of the cron job registration to use our standard
plugin registry.

Along the way we get rid of the 1.6 compatibility code to register cron
jobs. This also removes the undocumented way to register cron jobs using
the old GUI plugin module structure, which will make future refactorings
easier.

Change-Id: I5fd9ca7941c5b5c8ec8042f1806175ec5be7ea94