Make test_checks_sanity more specific
In the past this test was flaky because it was in
fact monitoring the CI build nodes. We have various
checks that monitor the systems resources; those would
occasionally be not OK (for instance the fetcher usage
being above the default levels or cached plugins
exceeding their expected execution time).

This was not the purpose of the test.

Change-Id: I9fb1a449e1c3fae756b5cfeda6e2cf8f9122c319