17381 FIX Agent Updater: Prevent initial reinstallation on Linux/Solaris
Previously, the agent updater relied on a previous update to know its own installed
agent hash. This lead to the situation that the agent updater always downloaded the
agent after initial registration, even if it was idential to the one already installed.

Now, the agent updater makes use of static information about the installed agent hash
that already comes with the installation of the agent package.

Change-Id: I5785c001ef20079edbc823bcaa00435e1d9400ef