gui-crawl: enforce completion of crawl-tasks after exiting loop.
Towards the end of the gui-crawl tests,
crawl tasks are not created when new URLs are not found,
which results in exit of while loop. The loop also has the logic
to wait for task-completion, which is skipped upon exiting.

Thus, sometimes leading to 'orphaned' tasks and hanging test-runs.

+ minor: improved comments

CMK-20093

Change-Id: If6108e9a42ca0c8343848a40345dada769b62de9