Agent bakery super_server: Move config files from /etc to our config
This is a preparation step for supporting the new single directory
deployment for agent packages.

By that, we want to step away from supporting /etc/check_mk as additional
configuration for baked agents. It will still be possible to deploy files
explicitly under /etc, since this is also part of the official Bakery API,
but we will stop to deploy any of our own config files there.

This change is compatible with the raw edition, where both paths
resolve to /etc/check_mk equally.

CMK-19980

Change-Id: I994f97dacddf238da544f9612c9715a2c3a61432