17250 FIX notifications: add 'emergency' as a Pushover priority option
For notification plugins, the priority option was missing from
available options in the rest-api. This werk fixes this issue.
However, this means that the schema also changes to accomodate
for this new option.

Previously, you could set one of the priority options like
this.

'''
"priority": "low"  # "lowest", "high", "normal"

'''

However, the "emergency" priority level adds a few more options,
so now, the schema looks like this

'''
"priority": {
    "level": "low"	# "lowest", "high", "normal"
}
'''

or

'''
"prioriy": {
    "level": "emergency",
	"retry": 60,
	"expire": 360,
	"receipt": "your receipt"
}
'''

Change-Id: Ibee77262c9730ebb638d355cc54f317a080fcda2