17358 FIX Disable Refresh on Background Jobs Overview Page
Before this fix, the site would reload every 0.8 seconds on the
background jobs overview page. This was problematic because the user
couldn't perform any actions during this small time window. Instead,
they would jump to the top of the overview page on each refresh.

This change removes the auto refresh functionality. So, now it is up to
the user to manually refresh when they want to see the current status
of a job.

SUP-20657
SUP-20821
SUP-21069

The actual change was made in a prior commit - the ref is linked below.

Ref: ac2daf4b8d4a6decc583a44f6306cc3982426aa1
Change-Id: I3e825409129cf85cb4d61a24befc1970265efbb3