Quick Setup: multiple actions per stage
This change allows for more than one ‘go to the next stage’ action per
stage. Each action is identified with an ID that is sent to the backend
when a stage needs to be validated. This enables actions to be created
in the backend that skip certain validations (in general, those that
take too much time).

The backend part of this change is CMK-20105

CMK-20174

Change-Id: Ic82ee9f9e77c40b21cb61e43299564a4a0e29f45