Move UI job scheduling from cron to dedicated process
This change moves the UI job triggering from the cron job to a dedicated
process. This change already reduces the overhead of job creation and
startup which was so far done once a minute, which is always nice.

However, more important are the things which are enabled by this:

a) Individual jobs can be controlled in a more fine-grained way, e.g. by
   allowing a per-job schedule without each job having to implement this
   in a slightly different way.
b) Other forms of jobs can be introduced, e.g. longer running jobs.
   Which reduces the overhead for initialization. So far, when a job
   typically takes longer to execute, we had to use background jobs, which
   needed one invocation of a job per iteration, which comes with some
   overhead. We can introduce a more lightweight approach for such jobs
   now.

Change-Id: Ib34d9e2ad4c6440fe9f85ccf10a3b850addb626e