17252 FIX activate_changes: sync folder deletion with remote site
When deleting a folder (with a host) on the central site,
that was being monitored on a remote site, the host
on the remote site would be deleted but the folder would
remain.

This werk fixes this issue by deleting both the host and
the folder on the remote site.

SUP-17381

Change-Id: I7e1b83cb8fa1777c4030f3118b319d200b3bd8ce