Improve background job start error handling
Get rid of the BackgroundJobAlreadyRunning and transport the information
to the caller using a result object instead.

This way we get rid of search_index job start calls regularly producing
"exception" traces when the job is already running. For this job, we
don't want to highlight this case, which is easier with this new
interface.

Change-Id: I970508a49712fc0f685767e7aeddfbd0a45ae3ce