17312 FIX Ruleset API: Fix rendering of configured FileUpload rules
Previously, after configuring a rule using the `cmk.rulesets.v1.form_specs.FileUpload` FormSpec, the overview of configured rules would display
```
The value of this rule is not valid.
```
Due to a bug in the rendering functionality this error is always shown even if the value is valid.
This has been fixed to show the selected file name instead.

SUP-21351

Change-Id: Icdd9a4f861c3036f1e25aa13d2a38718c2bcf545