Tracing: Add option to set resource attributes via config
It is now possible to save a mapping of resource attributes to the
file "etc/omd/tracing_resource_attributes.json". This file will be
picked up by our services and used to set the resource attributes
of the instrumented services.

We will use this in the CI to enrich our traces with some static meta
information regarding the runtime environment.

CMK-20270

Change-Id: I9e4859e2fc2a84e673837e517ab35614b88d0c44