17367 FIX Determine systemd enabled state from status section
You may be affected by this werk in case you're using systemd service templates in combination with `indirect` units.

Previous to this werk, the enabled state of your service was determined by looking at its unit file which may have indicated `indirect`.
This will result in the service beeing considered as `disabled` and therefore won't affect the status of the (Checkmk) service in case you've been using the summary check.

As the agent section `systemd_units` delivers also the `status` of a systemd service, the check can derive the enabled status from this information.
Since this information was not always present from the agent, a fall back to use the information `list-unit-files` is still in place.

Change-Id: I21dd250b6db6603bf3f1dc775f86f7522ddc767b
JIRA-Ref: SUP-21093