No longer prevent fetching during discovery
Previously we prevented fetching during discovery, since two
plugins would send data only once and that data had to be run
through the checking process.
This is not an issue anymore:
 * The logwatch agent plugin sends evey encountered log message
   for a set amount of time. As long as this amount is larger
   than or equal to a checking cycle, we're good.
 * `mk_inventory` produces persited sections, these are stored
   on the site anyway.

For now, lets only change this flag to have an easily revertable
change.

This might obsolete CMK-13546.
The proper cleanup of the remains of the hack is subject of CMK-17261.

Change-Id: Ib61668c68c9c8452520a4946dcce5b5c6b10d47b