17271 FIX Opsgenie: Changes to SSL validation
With [Werk #17270](https://checkmk.com/werk/17270) the Checkmk CA trust store
was supposed to be used with custom domains. Instead, it was only used for the
official domains. This shouldn't negatively impact anyone using the official
domains, but prevented those with custom domains from using the trust store.

With this werk, the trust store is now used for all domains. Further, an option
to disable SSL validation entirely has been added.

Change-Id: I7c238f647db27a56dc866778fbab191b0d4c33c6
JIRA-Ref: SUP-20550