Removed superfluous boolean parameter
A log entry is about a service if and only if the service description is
not empty. Or put another way: Host entries always have an empty service
description, and service entries always have a non-empty service
description.  If that's not the case, the monitoring history is broken
and the statehist computation will yield nonsense, anyway.

Nice side effect: This removes an anti-pattern (boolean parameter).

Change-Id: Ie64076f4737144caa62c2dcfc837c9b182d5b77b