docker sit: Improve container log evaluation
This change adds a log property for easy acces to the container logs.

Also, instead of only checking for a success message, we do now also
check for a known failure message during the container startup. This
will exit earlier in case of a failure, instead of waiting for the
complete timeout period.

CMK-20255

Change-Id: I5f58e754bad728ccadf5db53631e4882d03908d2