Lower cardinality of run_process span name
According to
https://github.com/open-telemetry/opentelemetry-specification/blob/main/specification/trace/api.md#span
span snames should be low cardinality.

This change reduces the cardinality of the span name for the run_process
background job, from
`run_process[activate-changes-scheduler-c3178f31-4928-47fa-96ca-5300e4d6ee22]`
to `run_process[activate-changes-scheduler]`. I am not sure if we should
reduce it even further to `run_process`. Looking at the Jaeger UI it
feels like a step back in terms of usability. We are still learning...

Change-Id: Ibd259fa1c63900f59e2e5facbcceb32995631731