quick_setup: fix bug in overview mode
The overview mode runs all stage validations, once after each other,
when saving.

Before this change, when running the validation for the
current stage, the FormSpec validation was really only running for the
current stage, but for running the custom validation all FormSpec data
has been parsed, even if not validated.

This meant when for example in Stage 2 there was a FormSpec validation error
the custom validation for Stage 1 crashed.

Now we are only parsing the data up until the current stage.

Now the custom validators also only validate the current stage data.

Change-Id: If9c09190605d6ddf4ebb6882473d576d4e1fc7e0