automation_helper: add initial daemon service
This change mainly deals with wiring the gunicorn application into
a checkmk site. The actual handling of automation command execution
will be picked up in subsequent changes.

The run command (defined in rc.d/) was set to 55 so that it runs before
the cmk-ui-jobs service (60) and apache web service (85). Both are known
services that rely on the automation commands.

Although `gunicorn` and `fastapi` were already available as indirect
dependencies, we need to explicitly define them in the `Pipfile` as
cmk project packages.

Change-Id: Ib89d618682da3e699f9e401928b9f87660f4a50c