automation_helper: fix unix sock file permission
Set the file permission of the socket generated by gunicorn to
umask=0077. This gives all permissions to the user and no permissions to
group or other. This is the same permission set for the dcd.sock.

In addition, we need to register the automation-helper as a new init
script so that the integration tests pass.

Change-Id: I9df96487a265be80ab104072efecb3ad39d59fff