gui-e2e: increase timeouts from 30 secs to 120 secs.
Tests expect certain web-elements to occur in the UI within 30 secs
of an interaction. During high CI loads, this might not be possible.

Thus, increasing timeout thresholds as the tests do not validate performance of the site.

+ consistent usage of timeout in gui-e2e.

NOTE: affects duration of the testsuite run
(especially when failing tests are present)

CMK-20395

Change-Id: I459ed7c55625be5a4a033c4c73f91ec04b1c0d4e