gui-crawl: Reduce memory required to end a test-run.
Earlier,
If URLs were not found for 100 iterations, then test-run was exited.
This took ~8 mins on a local system. On a busy CI node, the time taken
could be scaled by a fator of 3.

Now,
If URLs are not found for 25 iterations, then exit the test-run.

Removed `time.sleep(rate_of_tasks)`.
processing the tasks as a TaskGroup ensures that URLs are found faster
than the tasks being created.

CMK-20093

Change-Id: I0fd9e9b4023d29e7c1a1cf66853ba4ab086e4b1d