17382 FIX Agent Controller (Windows): Stop querying the agent receiver on agent run
On every agent run, the agent controller is queried for its status as a data basis
(among other sources) for the _Check\_MK Agent_ service.
In the Linux agent, the call to `cmk-agent-ctl status` is prevented to contact the
Check\_MK site (as it's normally done) by adding the `--no-query-remote` option,
but this was missing in the Windows agent.

SUP-19858

Change-Id: Ifef65cfa7e801ef34e1fc8f58f30bafa7d702b0c