17335 Fix rules not matching non-central cmk/site labels
Due to the order of our host label resolution, we overwrote any explicit
`cmk/site` labels with the central site label. This caused rules to not
match hosts with non-central site labels.

We change the order of the label resolution to first check for the
built-in before overwriting them with the explicit labels.

Since the builtin labels currently only contains the `cmk/site` label,
this change has no other effect, but this change itself affects the
monitoring already substantially.

SUP-20807

Change-Id: I6f645febb3dc50d530ad1bf726a917579cc7d259