15348 FIX check_mk_agent.solaris: Hide statgrab output in case of timeout
This is a followup to [Werk #15284](https://checkmk.com/werk/15284)
which added a timeout to the `statgrab` command.

In certain cases it's possible that the `statgrab` command times out in
the middle of outputting data. This can lead to the problem, that some
output is missing for a certain disk. For example `disk_name` is shown,
but `write_bytes` is missing, which leads to a crash in the check
plugin.

With this Werk, the whole `statgrab` output will be rejected in case of
a timeout, as we can not know if the output of `statgrab` is complete.

SUP-21231

Change-Id: I610e346629de32cbba962d0a3320eb7e0525f8c4