automations: move reload flag to execute method
The automation integration tests do not make use of the `automations`
singleton directly. Instead the automations are called directly from the
command line. So the flag in the module scope is never picked up. By
placing the flag in the method directly, we can be sure that it is
picked up. However, the flag should be removed from the method signature
post migration.

Change-Id: I394a49c871ac276832182202830feda4af183991