Add capability to start UI jobs in threads
This change is a preparation to migrate some background jobs launched by
the so-called UI jobs from subprocesses to threads to reduce the job
initialization overhead.

Background:

We have 5 background jobs which run in a one-minute interval in central
sites of distributed setups. Their process initialization causes a
constant load.

This load is way higher than it was in 2.3. This is because the
background jobs have been changed from fork to spawn model in 2.4 to
have a cleaner state of the application when being started. This comes
with a significantly higher overhead of process initialization.

A few days ago we introduced the cmk-ui-jobs process which now schedules
the regular UI jobs. In this continously running process we have more
freedom to manage jobs. We can now start independent threads and keep
them running for longer. This removes the need for such background jobs.

Change-Id: I15144c110ac6885f7d4e77d2972c1e185713a488