systemd: better message
This change is meant to help users with an agent section such as:
```
<<<systemd_units>>>
[list-unit-files]
thermald.service disabled enabled
[status]
[all]
crio.service not-found inactive dead crio.service
```

Note, that systemctl could not properly load the crio.service unit
definition:
```sh
$ systemctl status crio
Unit crio.service could not be found.
```
However, the crio.service is still considered in memory, and will returned by
`systemctl --all --type service --type socket`.

On the other hand, the thermald.service is properly loaded, but has been
stopped and disabled:
```sh
$ systemctl status thermald
○ thermald.service - Thermal Daemon Service
     Loaded: loaded (/usr/lib/systemd/system/thermald.service; disabled; preset: enabled)
     Active: inactive (dead)
```
In Checkmk, the `crio.service` is discovered and `Status: inactive` is
reported. The `thermald.service` will not be discovered and Checkmk
reports `Unit not found`.

In this change:
* Add a debugging command to help users identify why a unit might be
  loaded, but won't be discovered by Checkmk.
* `Service` is replaced by `Unit`, which is more accurate.
* Updates the manpages, such that the discovery section correctly shows
  up on checkmk.com
* Adds some helpful information about the difference between
  'systemctl --all' and a loaded unit file.

CMK-21435

Change-Id: I12a981c104d99721cacd6a93853df2c93080fcc1