automation_helper: add support for observer
This observer context manager receives a sequence of schedules and
builds handlers based on the spec. With theses built handlers, the
observer then listens to changes to the filesystem and handle said
events with the handlers that the observer built.

Since we are dealing with a spec, we can now support all our watching
requirements from within one observer. This also makes it easier to
improve the implementation in the future by reading in the spec from a
configuration file, validating it in pydantic, and then injecting it
into the observer.

CMK-20432

Change-Id: I6abd71afcd9f51adaf2b6ab9a69541aeb03d41e6