Reland "17401 Performance metrics for "Check_MK Discovery" and "Check_MK HW/SW Inventory""
This reverts commit be1badd1e1ea095b5f9e70dff63c52a1c7551a0e.

The patterns of the nested contexts and the handling of the results
have been aligned to avoid the `UnboundLocalError`.

Note that the need to carefully prevent this error is a result
of the fact that we're using an exception-swallowing context
manager with a non-trivial body -- leading to potentially
unbound locals in the code after the 'with' block.

Change-Id: I2da86fc5dea9497bb4a72339d9c91ee2d971d46f