testlib.repo: Cache repository helper functions
The repository related helper functions are relatively costly,
given that they are run repeatedly from many different tests.
As the output of those functions is more or less static during a
test run, this change adds caching so we do not have to execute
the external commands multiple times.

CMK-20561

Change-Id: I6a2070ea8de4a2120e5bcfb404cdfd14b049cc61