systemd_units: Fix time since last status change
Previously, the time since the last status change was extracted with the assumption that the location of this information in the `systemctl status` is static. However, changes like adding a systemd drop-in file could shift the lines, causing the time to not be found.
Now the relevant location is identified dynamically to ensure it is always output correctly.

SUP-21456

Change-Id: I1266cbea8c699e7ebaa29abd5f6df3cf3e2360f7