Switch pagetype updaters to standard plugins, drop separate registry
Note that this changes the way these updaters work: Before, one could register
an updater for either one of the concrete implementations (eg. `ForecastGraph`)
or for one of the intermediate, abstract classes (eg. `OverridableContainer`).
With this change, the latter won't have any effect, since
`self._updater.target_type.load_raw()` won't return anything.

At the moment, we have no need for updaters targeting one of the abstract
classes. Should this change, we will have to re-think this design. On the other
hand, this change migrates the pagetype updaters to a standard plugin structure
and it paves the way for pre-update actions, see the subsequent changes.

Pre-update actions don't work nicely in combination with updaters targeting the
abstract classes. During the actual update process, you can simply first call
the updaters targeting the abstract classes and the updaters targeting the
concrete implementations (since the latter might rely on changes done by the
former). However, during the pre-actions, this approach does not work because
the changes are not persisted. So, a later pre-update step cannot rely on
changes done by a former step.

SUP-21640

Change-Id: Ic9929666a9749cec16362cc5d093f15e8eaa37d7