Fix test_auto_registration_workflow_remote flake
The SyncRemoteSitesJob syncs site changes from remote sites to the
central site.

When e.g. DiscoverRegisteredHostsJob runs on the remote site, creates a
change and activates it, it may happen that SyncRemoteSitesJob runs
before the activation finishes, which breaks the logic of
DiscoverRegisteredHostsJob.

To mitigate this, the SyncRemoteSitesJob now only processes changes
which are 10 minutes or older.

Change-Id: I694dca359047ad9e861fe990e0a6ceadc179ecac