Service discovery: remove reclassification of changed services due to service parameters
In 2.3 we introduced the detection of changed services in the backend,
for changed service labels and service parameters.

Since we had no way of displaying this information in the frontend we
decided to reclassify services that were changed due to service
parameters as unchanged.

With the new concept of displaying this information we can get rid of
the reclassification.

CMK-20313

Change-Id: I8b7c19784d47097804e4e876280b5ab8a3beebf1