Fix agent bi and formspecs
AgentBI used to get its config via stdin. With the migration to
formspecs and server-side-calls the back-translation of secrets stopped
working since we don't expose the real secrets to the API.

With this change we send the passwords on the commandline and the rest
via stdin. We merge the config and the password spec on runtime. The
agent config supports an arbitrary long list of configs so we must
expect an arbitrary number of passwords on the commandline.

If there are too many passwords we get into trouble. According to my
research the limit for the commandline is variable from os to os and is
even configurable. On my system the agent was able to handle a
commandline with more than 4000 bytes. I don't expect that to happen in
the real world though.

Change-Id: I0b05d33bc096931f6c8a009e52b52ae38d2ae2cf