17274 FIX password: owner/owned_by fields replaced with editable_by
The request and response schemas for password endpoints use a mixture of `owner`
and `owned_by` in their fields. To consolidate this, and make it consistent
with the UI, a new field `editable_by` will be added with this werk. The old
fields still work, but will be removed in version 2.5.

In addition, users with `wato.edit_all_passwords` permission will be able to
create/update passwords for contact groups they are not members of.

Change-Id: Ia945ca79a69e838d1dbf332bb9268071b4bba2bd
JIRA-Ref: SUP-20743