17402 Improve "Fetch intervals for SNMP sections" ruleset
This extends the ruleset "Fetch intervals for SNMP sections".
Users can now not only set up the rule for _all_ or _one_ section, but rather for any collection of sections.
Additionally users can configure _"Fetch data every time"_, which corresponds to
the default behavior (as if no rule was configured at all).

Previously users could only configure a value of 1 minute, which is **not** the same as the default.

Change-Id: Id417524a8188809c35246721a23420251646d724