Licensing: Allow licensing of sites before upgrade from CEE
Previously not all the relevant license information was available when upgrading an Enterprise to another edition, necessitating [werk #16683](https://checkmk.com/werk/16683).
With this werk, it is now possible to apply a license compatible with the upgraded version before the upgrade and keep all the information during the upgrade (the license has to be compatible to the installed edition as well).
If the applied license is incompatible with the upgraded version, the license state will still be reset.

To check the compatibility see [werk #17310](https://checkmk.com/werk/17310).

CMK-20427

Change-Id: If43e5a2502199aad667e58db2427b0dc7ede2e39