17417 Fix host_group_config: update endpoint family description"
The host_group_config endpoint family has been updated to provide a more
informational description for the members field.  This field is returned
in all of our responses but in this case, it was not clear that the members
field would not contain the host group members.  In order to get the host group
members, you must call the relevant 'monitor' endpoints.

SUP-20864

Change-Id: I1075d720b6c926538813c301091aabc5cefd7731