17003 REST API: change "Show hosts for specific condition" endpoint method from GET to POST
The REST API endpoint "Show hosts for specific condition" was only accessible via the GET method. This caused a limitation of the maximum size of query parameters that users could send in a request.
A common practice to bypass this limitation is to use a POST request which in this case allows users to send requests of larger and more complex queries.

With this werk the REST API endpoint "Show hosts for specific condition" is now accessible via a POST request, taking all the formerly possible parameters as part of its request body.

CMK-20251

Change-Id: I2cb8c6f16c11ac7ecdfe0d6b83fb5ed891299114