Only show required when dict. el. is non-optional
This was the result of a little brainstorming with UX.

We don't want to show a (required) next to an optional dictionary
element title even if the child form spec has a validator
which would fail if no value is defined. It would look as if the
optional key needs to be checked.

We open ourself to the inconsistency that if the child form spec does
not have a label, we miss rendering the (required) completely.

This however is preferable to the alternatives of always showing the
required even if no label is present.

CMK-19938

Change-Id: Id47c11b0842cf7b7037641c081b53643b267edab