17368 FIX Support Prism Central
You may be affected by this werk if you're having prism central and prism element instances in your monitoring.
A possible traceback of the special agent may look like:
```
agent_prism: HTTP error: 412 Client Error: PRECONDITION FAILED for url
```
The reason therefore is, that the following endpoints cannot be queried on the prism central:

* `protection_domains`
* `remote_support`
* `ha`
* `hosts_networks`

The agent now detects if it is a central instance and skips querying those endpoints.

Change-Id: I24ea11041336c9b84bcdb80e5106151d0a496f88
JIRA-Ref: SUP-21383